Fixed mistake in code which caused creatures doing nothing to wander the dungeon instead of going to sleep. Also replaced the non-existing GROUP job with BARRACK for skeletons.

+++ keeperfx/src/creature_states_train.c
@@ -143,7 +143,7 @@ struct Thing *get_creature_in_training_room_which_could_accept_partner(struct Ro
         cctrl = creature_control_get_from_thing(thing);
         if (!creature_control_exists(cctrl))
         {
-            ERRORLOG("Jump to invalid creature %ld detected",i);
+            ERRORLOG("Jump to invalid creature %d detected",(int)i);
             break;
         }
         i = cctrl->next_in_room;
